Xdebug with Docker.
I have finished the xdebug with docker. I also inluded how to set it up inside PHPstrom since that in the IDE that I have been using.  I made it in its own folder/image so that it cant be confused with production. It is very handy to have set up if you are making any changes. Let me know if you have any questions or problems.