Add in the auto_configure.php to root
The devtools for restore / backup require the auto_configure.php to be
in the /root/ folder.  The auto_configure.php only gets copied to root
via the openemr.sh file if its an easy dev docker setup.  However this
messes up the ci installations (such as inferno) which uses the
restore/backup in a non-easy dev type setup (where the codebase is not
shared).

Just copying auto_configure.php to root solves that issue.